requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
arabic_reshaper=2.1.4=pyhd8ed1ab_0
asgiref=3.5.2=py310h06a4308_0
asn1crypto=1.5.1=pyhd8ed1ab_0
asttokens=2.0.5=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
bcrypt=3.2.0=py310h5eee18b_1
biopython=1.78=py310h7f8727e_0
blas=1.0=openblas
bottleneck=1.3.5=py310ha9d4c09_0
brotli=1.0.9=h166bdaf_7
brotli-bin=1.0.9=h166bdaf_7
brotlipy=0.7.0=py310h5764c6d_1004
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.12.7=ha878542_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.0=py310h0fdd8cc_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=unix_pyhd8ed1ab_2
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.2=py310h06a4308_0
configparser=5.3.0=pyhd8ed1ab_0
cryptography=3.4.8=py310h685ca39_1
cssselect2=0.2.1=pyh9f0ad1d_1
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py310h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
django=4.1=py310h06a4308_0
django-crispy-forms=1.14.0=pyhd8ed1ab_0
django-filter=22.1=pyhd8ed1ab_0
django-widget-tweaks=1.4.5=py_0
entrypoints=0.4=py310h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.9=h6a678d5_0
faker=8.8.1=pyhd3eb1b0_1
fftw=3.3.9=h27cfd23_1
fontconfig=2.14.1=h52c9d5c_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
future=0.18.3=pyhd8ed1ab_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=he621ea3_2
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
html5lib=1.1=pyh9f0ad1d_0
icu=58.2=he6710b0_3
idna=3.4=pyhd8ed1ab_0
ipykernel=6.19.2=py310h2f386ee_0
ipython=8.7.0=py310h06a4308_0
jedi=0.18.1=py310h06a4308_1
jpeg=9e=h7f8727e_0
jupyter_client=7.4.8=py310h06a4308_0
jupyter_core=5.1.1=py310h06a4308_0
kiwisolver=1.4.4=py310h6a678d5_0
krb5=1.19.2=hac12032_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libblas=3.9.0=15_linux64_openblas
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libcblas=3.9.0=15_linux64_openblas
libclang=10.0.1=default_hb85057a_2
libedit=3.1.20221030=h5eee18b_0
libevent=2.1.12=h8f2d780_0
libffi=3.4.2=h6a678d5_6
libgcc=7.2.0=h69d50b8_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=12.2.0=h69a702a_19
libgfortran5=12.2.0=h337968e_19
libgomp=11.2.0=h1234567_1
liblapack=3.9.0=15_linux64_openblas
libllvm10=10.0.1=hbcb73fb_5
libopenblas=0.3.20=pthreads_h78a6416_0
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.2.0=h2818925_1
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lxml=4.9.1=py310h1edc446_0
lz4-c=1.9.4=h6a678d5_0
matplotlib=3.5.3=py310hff52083_2
matplotlib-base=3.5.3=py310hf590b9c_0
matplotlib-inline=0.1.6=py310h06a4308_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.6=py310h06a4308_0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numexpr=2.8.3=py310h757a811_0
numpy=1.22.3=py310h4ef5377_2
openssl=1.1.1s=h7f8727e_0
oscrypto=1.2.1=pyhd3deb0d_0
packaging=22.0=py310h06a4308_0
pandas=1.4.3=py310h6a678d5_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py310hace64e9_1
pip=22.3.1=py310h06a4308_0
platformdirs=2.5.2=py310h06a4308_0
ply=3.11=py_1
prompt-toolkit=3.0.36=py310h06a4308_0
psutil=5.9.0=py310h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.11.2=pyhd3eb1b0_0
pyhanko=0.16.0=pyhd8ed1ab_0
pyhanko-certvalidator=0.19.8=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pypdf3=1.0.6=pyhd8ed1ab_0
pyqt=5.15.7=py310h6a678d5_1
pyqt5-sip=12.11.0=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.8=h7a1cb2a_1
python-bidi=0.4.2=py_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2022.7=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytz=2022.1=py310h06a4308_0
pytz-deprecation-shim=0.1.0.post0=py310hff52083_3
pyyaml=6.0=py310h5764c6d_4
pyzmq=23.2.0=py310h6a678d5_0
qrcode=7.3.1=pyhd8ed1ab_0
qt=5.15.9=h06a4308_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
reportlab=3.5.67=py310h0edfc0f_1
requests=2.28.2=pyhd8ed1ab_0
scipy=1.7.3=py310hac523dd_2
seaborn=0.11.2=pyhd3eb1b0_0
setuptools=65.6.3=py310h06a4308_0
sip=6.6.2=py310h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.40.1=h5082296_0
sqlitebrowser=3.8.0=0
sqlparse=0.4.3=py310h06a4308_0
stack_data=0.2.0=pyhd3eb1b0_0
svglib=1.5.1=pyhd8ed1ab_0
text-unidecode=1.3=pyhd3eb1b0_0
tinycss2=1.2.1=pyhd8ed1ab_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd8ed1ab_0
tornado=6.2=py310h5eee18b_0
tqdm=4.64.1=pyhd8ed1ab_0
traitlets=5.7.1=py310h06a4308_0
tzdata=2022g=h04d1e81_0
tzlocal=4.2=py310hff52083_2
uritools=4.0.1=pyhd8ed1ab_0
urllib3=1.26.14=pyhd8ed1ab_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py_1
wheel=0.37.1=pyhd3eb1b0_0
xhtml2pdf=0.2.8=py310hff52083_0
xz=5.2.8=h5eee18b_0
yaml=0.2.5=h7f98852_2
zeromq=4.3.4=h2531618_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
